+++ examples/wildcards_monitor.py
@@ -13,11 +13,11 @@ def on_metric(client, ud, msg):
     print("Metric@", str(msg.topic), "=", msg.payload.value)
 
 client.on_message = on_metric
-client.subscribe(Topic(payload_type=Metric, context=("#",)))  # example/v1/_Metric/#
+client.subscribe(Topic(payload_type=Metric, node_id="+", context=("#",)))  # example/v1/_Metric/+/#
 
 # publish a couple for demo
-client.publish(Topic(payload_type=Metric, context=("line-1","motor-7","speed")), Metric(123.4))
-client.publish(Topic(payload_type=Metric, context=("line-2","oven","temp")), Metric(456.7))
+client.publish(Topic(payload_type=Metric, node_id="metric-tap", context=("line-1","motor-7","speed")), Metric(123.4))
+client.publish(Topic(payload_type=Metric, node_id="metric-tap", context=("line-2","oven","temp")), Metric(456.7))
 
 time.sleep(1)
 client.loop_stop(); client.disconnect()